requirements.txt
black==22.12.0
fastapi==0.89.1
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-pubsub==2.14.0
ipykernel==6.20.2
ipython==8.9.0
pydantic==1.10.4
pytest==7.2.1
requests==2.28.2
uvicorn==0.20.0